requirements.txt
sys
sqlite3
random.randint
PyQt5.QtWidgets.QWidget
PyQt5.QtWidgets.QApplication
PyQt5.uic